Add Probability and Distribution Project
In this project, data collected from the failure of three different models is analyzed. With all the failure information collected, a distribution function is sought that best fits the actual failure probability of each one, in order to implement safety systems that reduce the probability of these components being damaged.